requirements.txt
Flask==2.3.3
flask-cors==4.0.0
langgraph>=0.6.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langfuse>=3.0.0
python-dotenv>=1.0.0
